Conflicts:
	ECom.Domain/Aggregates/Order/OrderAggregate.cs
	ECom.Domain/ECom.Domain.csproj
	ECom.Infrustructure/ECom.Infrastructure.csproj
	ECom.Messages/Contracts.cs
	EComWithCQRS.sln

+++ ECom.Domain/IDomainIdentityGenerator.cs
@@ -10,6 +10,6 @@ namespace ECom.Domain
 	/// </summary>
 	public interface IDomainIdentityGenerator
 	{
-		int GenerateNewId();
+		long GenerateNewId();
 	}
 }
